Add CSS-based line numbering to code blocks
This might be fragile, but so far seems to work surprisingly well! As noted
it is based heavily on the PrismJS line numbering, and only works for our
pre elements that are not line-wrapped!
