+++ src/index-cs.tsx
@@ -7,7 +7,7 @@ import {Provider} from "react-redux";
 import {store} from "./app/state/store";
 import {IsaacApp} from './app/components/navigation/IsaacApp';
 import {printAsciiArtLogoToConsole} from "./app/services/easterEggs/csLogoAsciiArt";
-import * as highlightJsConfig from "./app/services/highlightJsConfig";
+import * as highlightJsConfig from "./app/services/highlightJs";
 
 printAsciiArtLogoToConsole();
 highlightJsConfig.registerLanguages();
