+++ src/scss/cs/isaac.scss
@@ -91,6 +91,7 @@ $tertiary-font: 'arconrounded-regular', sans-serif;
 @import "button";
 @import "logo";
 @import "featured-profile";
+@import "../common/code";
 
 // Navigation
 @import "header";
