+++ src/scss/phy/isaac.scss
@@ -110,6 +110,7 @@ $tertiary-font: 'arconrounded-regular', sans-serif;
 @import "logo";
 @import "cards";
 @import "featured-profile";
+@import "../common/code";
 
 // Navigation
 @import "../common/navigation-bar";
